Initial code transition from single-account methodology
This has a good amount, but not all changes that are being made first in the internal dap-configuration repo, used mainly to test this as we transition. I'll likely have to make another commit once I have time to synchronize the changes made to dap-configuration and clean them up so they can be public.